Removed prody/dynamics/solvate.py and added all SAXS related functions into saxs.py in the same directory
