+++ prody/dynamics/__init__.py
@@ -271,3 +271,7 @@ __all__.extend(vmdfile.__all__)
 from . import exanm
 from .exanm import *
 __all__.extend(exanm.__all__)
+
+from . import saxs
+from .saxs import *
+__all__.extend(saxs.__all__)
